+++ CHANGELOG.md
@@ -1,2 +1,8 @@
 ## v0.0.1
-Initial release
+- Initial release
+
+## v0.0.2
+- Extension activation fixes
+
+## v0.0.3
+- Experimental "TEAL: Install/Update Tools" command